Add the function that add new task
