+++ src/index.js
@@ -1,45 +1,21 @@
-/* eslint-disable max-classes-per-file */
 /* eslint-disable no-underscore-dangle */
 import './style.css';
+import currentTasks from './modules/Tasks';
+import add from './modules/add.js';
+import { addBtn, newTask, tasksContainer } from './modules/TaskElements.js';
 
-const tasksContainer = document.querySelector('.list-tasks');
+//add new task
 
-class Task {
-  constructor(description, completed, index) {
-    this._description = description;
-    this._completed = completed;
-    this._index = index;
+newTask.addEventListener('keypress', function (e) {
+  if (e.key === 'Enter') {
+    const task = add(e);
+    currentTasks.add(task);
   }
-}
+});
 
-class Tasks {
-  constructor() {
-    this._tasks = [];
-  }
-
-  add = (task) => {
-    this._tasks.push(task);
-  };
-
-  display = () => {
-    this._tasks.forEach((task) => {
-      const taskItem = document.createElement('li');
-      taskItem.innerHTML = `
-      <input type="checkbox" id="task-${task._index}" name="task-${task._index}" value="Bike">
-      <label for="task-${task._index}">${task._description}</label><br>
-      `;
-      tasksContainer.appendChild(taskItem);
-    });
-  };
-}
-
-const firstTask = new Task('wash dishes', false, 0);
-const secondTask = new Task('fix car', false, 1);
-const thirdTask = new Task('clean the house', false, 2);
-
-const currentTasks = new Tasks();
-currentTasks.add(firstTask);
-currentTasks.add(secondTask);
-currentTasks.add(thirdTask);
+addBtn.addEventListener('click', (e) => {
+  const task = add(e);
+  currentTasks.add(task);
+});
 
 currentTasks.display();
