+++ src/modules/TaskElements.js
@@ -1,0 +1,4 @@
+export const addBtn = document.querySelector('.fa-arrow-right');
+export const clearTasksBtn = document.querySelector('.clear-tasks');
+export const newTask = document.getElementById('new-task');
+export const tasksContainer = document.querySelector('.list-tasks');
